requirements.txt
aiogram==3.0.0b6
redis==4.4.2
loguru==0.6.0
pydantic==1.10.4
python-dotenv==0.21.1
nextcloud-api-wrapper==0.2.3
Telethon==1.26.1
tgcrypto==1.2.5